rename and clean up for common code
